refactor: 회원가입 양식 이메일 검증 기능 추가
