+++ src/main/java/roomescape/presentation/dto/request/SignUpRequestDto.java
@@ -1,9 +1,10 @@
-package roomescape.presentation.dto;
+package roomescape.presentation.dto.request;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
 
 public record SignUpRequestDto(
-        @NotEmpty String email,
+        @NotEmpty @Email String email,
         @NotEmpty String password,
         @NotEmpty String name
 ) {
